document_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ingestion"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai",
    "psycopg2-binary",
    "pgvector",
    "PyMuPDF",
    "python-docx",
    "python-pptx",
    "openpyxl",
    "tiktoken",
    "python-dotenv",
    "pyyaml",
    "requests",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
ingestion = ["prompts/*.yaml"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["ingestion*"]

[tool.black]
line-length = 79
target-version = ["py313"]

[tool.pylint.main]
source-roots = ["src", "tests"]
extension-pkg-allow-list = ["fitz"]

# R0801 (duplicate-code) disabled because processors are
# intentionally self-contained with duplicated conversion
# logic — plug/play architecture over DRY.
[tool.pylint."messages_control"]
disable = ["R0801"]

[tool.pylint.design]
max-attributes = 15
max-args = 7
max-positional-arguments = 7
max-locals = 20
max-returns = 7
max-statements = 60

[tool.pylint.format]
max-module-lines = 1700

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]

[tool.mypy]
mypy_path = "src:typings"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "mypy",
    "pylint",
    "black",
    "flake8",
]
